+++ ConcernsCaseWork/ConcernsCaseWork.CypressTests/cypress/e2e/casework-regression/create-case.cy.ts
@@ -5,6 +5,7 @@ import { CreateCasePage } from "cypress/pages/createCase/createCasePage";
 import CreateConcernPage from "cypress/pages/createCase/createConcernPage";
 import AddConcernDetailsPage from "cypress/pages/createCase/addConcernDetailsPage";
 import caseManagementPage from "cypress/pages/caseMangementPage";
+import concernsApi from "cypress/api/concernsApi";
 
 describe("Creating a case", () =>
 {
@@ -128,7 +129,7 @@ describe("Creating a case", () =>
         caseManagementPage
             .hasTrust("Ashton West End Primary Academy")
             .hasRiskToTrust("Red Plus")
-            .hasConcerns("Financial: Deficit")
+            .hasConcerns("Financial: Deficit", ["Red", "Amber"])
             .hasTerritory("North and UTC - North East")
             .hasIssue("This is an issue")
             .hasCurrentStatus("This is the current status")
@@ -136,6 +137,14 @@ describe("Creating a case", () =>
             .hasDeEscalationPoint("This is the de-escalation point")
             .hasNextSteps("This is the next steps")
             .hasCaseHistory("This is the case history");
+
+        Logger.Log("Verify the means of referral is set");
+        caseManagementPage.getCaseIDText().then((caseId) => {
+            concernsApi.get(parseInt(caseId))
+                .then(response => {
+                    expect(response[0].meansOfReferralId).to.eq(2);
+                });
+        });
     });
 
     it(("Should create a case with only required fields"), () => {
@@ -193,7 +202,7 @@ describe("Creating a case", () =>
         caseManagementPage
             .hasTrust("Ashton West End Primary Academy")
             .hasRiskToTrust("Red Plus")
-            .hasConcerns("Financial: Deficit")
+            .hasConcerns("Financial: Deficit", ["Red", "Amber"])
             .hasTerritory("North and UTC - North East")
             .hasIssue("This is an issue")
             .hasEmptyCurrentStatus()
@@ -201,5 +210,97 @@ describe("Creating a case", () =>
             .hasEmptyDeEscalationPoint()
             .hasEmptyNextSteps()
             .hasEmptyCaseHistory();
+    });
+
+    const searchTerm =
+		"Accrington St Christopher's Church Of England High School";
+	const searchTermForSchool = "school";
+
+    it("User searches for a valid Trust and selects it", () => {
+
+        createCasePage
+            .createCase()
+            .withTrustName(searchTerm)
+            .selectOption()
+            .confirmOption();
+
+		Logger.Log("Should display the Concern details of the specified Trust");
+		createConcernPage
+            .hasTrustSummaryDetails(searchTerm)
+            .cancel();
+	});
+
+    it('Should display an error if no trust is selected', () => {
+		createCasePage
+            .createCase()
+            .withTrustName("A")
+            .confirmOption()
+            .hasValidationError("Select a trust");
+    });
+
+	it('Should display a warning if too many results', () => {
+		createCasePage
+            .createCase()
+            .withTrustName(searchTermForSchool)
+            .shouldNotHaveVisibleLoader()
+            .hasTooManyResultsWarning("There are a large number of search results. Try a more specific search term.");
+    });
+
+    it("Should create additional concerns", () =>
+    {
+        Logger.Log("Create a case");
+        createCasePage
+            .createCase()
+            .withTrustName("Ashton West End Primary Academy")
+            .selectOption()
+            .confirmOption();
+
+        Logger.Log("Create a valid concern");
+        createConcernPage
+            .withConcernType("Irregularity")
+            .withSubConcernType("Irregularity: Suspected fraud")
+            .withRating("Red-Plus")
+            .withMeansOfRefferal("Internal")
+            .addConcern();
+
+        Logger.Log("Adding another concern during case creation");
+        createConcernPage
+            .addAnotherConcern()
+            .withConcernType("Governance and compliance")
+            .withSubConcernType("Governance and compliance: Compliance")
+            .withRating("Amber-Green")
+            .withMeansOfRefferal("External")
+            .addConcern()
+            .nextStep();
+
+        Logger.Log("Populate risk to trust");
+        addDetailsPage
+            .withRating("Red-Plus")
+            .nextStep();
+
+        Logger.Log("Populate territory");
+        addTerritoryPage
+            .withTerritory("North and UTC - North East")
+            .nextStep();
+
+        Logger.Log("Add concern details with valid text limit");
+        addConcernDetailsPage
+            .withIssue("This is an issue")
+            .createCase();
+
+        Logger.Log("Add another concern after case creation");
+        caseManagementPage.addAnotherConcern();
+
+        createConcernPage
+            .withConcernType("Irregularity")
+            .withSubConcernType("Irregularity: Irregularity")
+            .withRating("Red-Amber")
+            .withMeansOfRefferal("External")
+            .addConcern();
+
+        caseManagementPage
+            .hasConcerns("Irregularity: Suspected fraud", ["Red Plus"])
+            .hasConcerns("Governance and compliance: Compliance", ["Amber", "Green"])
+            .hasConcerns("Irregularity: Irregularity", ["Red", "Amber"]);
     });
 });