+++ ConcernsCaseWork/ConcernsCaseWork.CypressTests/cypress/e2e/casework-regression/edit-case.cy.ts
@@ -13,6 +13,8 @@ import EditCaseAimPage from "cypress/pages/createCase/editCaseAimPage";
 import EditDeEscalationPointPage from "cypress/pages/createCase/editDeescalationPointPage";
 import EditNextStepsPage from "cypress/pages/createCase/editNextStepsPage";
 import EditCaseHistoryPage from "cypress/pages/createCase/editCaseHistoryPage";
+import CaseManagementPage from "../../pages/caseMangementPage";
+import AddToCasePage from "../../pages/caseActions/addToCasePage";
 
 describe("Editing a case", () =>
 {
@@ -21,7 +23,6 @@ describe("Editing a case", () =>
     const addDetailsPage = new AddDetailsPage();
     const addTerritoryPage = new AddTerritoryPage();
     const addConcernDetailsPage = new AddConcernDetailsPage();
-    // const caseManagementPage = new CaseManagementPage();
     const editRiskToTrust = new EditRiskToTrustPage();
     const editDirectionOfTravel = new EditDirectionOfTravelPage();
     const editIssuePage = new EditIssuePage();
@@ -80,7 +81,7 @@ describe("Editing a case", () =>
         caseManagementPage
             .hasRiskToTrust("Red Plus")
             .hasDirectionOfTravel("Deteriorating")
-            .hasConcerns("Red")
+            .hasConcerns("Financial: Deficit", ["Red"])
             .hasTerritory("North and UTC - North East")
             .hasIssue("This is an issue");
 
@@ -110,7 +111,7 @@ describe("Editing a case", () =>
 
         addDetailsPage
             .hasRating("Red")
-            .withRating("Red-Plus")
+            .withRating("Amber-Green")
             .apply();
 
         Logger.Log("Edit a territory")
@@ -223,12 +224,11 @@ describe("Editing a case", () =>
             .withCaseHistory("New case history")
             .apply();
         
-
         Logger.Log("Verify detailes have been changed")
         caseManagementPage
             .hasRiskToTrust("Red")
             .hasDirectionOfTravel("Improving")
-            .hasConcerns("Red Plus")
+            .hasConcerns("Financial: Deficit", ["Amber", "Green"])
             .hasTerritory("North and UTC - North West")
             .hasIssue("New Issue")
             .hasCurrentStatus("New Status")
@@ -236,5 +236,14 @@ describe("Editing a case", () =>
             .hasDeEscalationPoint("New de-descalation point")  
             .hasNextSteps("New next step")
             .hasCaseHistory("New case history");
+    });
+
+    it("Should raise a validation error if do not select a case action", () => 
+    {
+        cy.basicCreateCase();
+
+        CaseManagementPage.getAddToCaseBtn().click();
+        AddToCasePage.getAddToCaseBtn().click();
+        AddToCasePage.hasValidationError("Please select an action to add.");
     });
 });