+++ ConcernsCaseWork/ConcernsCaseWork.CypressTests/cypress/pages/caseMangementPage.ts
@@ -186,6 +186,14 @@ class CaseManagementPage {
         return this;
     }
 
+    public addAnotherConcern(): this
+    {
+        Logger.Log("Adding another concern");
+        cy.getByTestId("add-additional-concern").click();
+
+        return this;
+    }
+
     public editRiskToTrust(): this {
         Logger.Log("Editing the risk to trust");
         this.getEditRiskToTrust().click();
@@ -248,6 +256,61 @@ class CaseManagementPage {
     {
         Logger.Log("Cannot edit the territory");
         this.getEditTerritory().should("not.exist");
+
+        return this;
+    }
+
+    public hasCaseOwner(value: string): this {
+        Logger.Log(`Has case owner ${value}`);
+
+        // Can be improved later
+        // Currently its driven by the casing of the email when the user logs in
+        // We can't control this, so safer to ignore case for now
+        cy.getByTestId("case owner_field").contains(value, { matchCase: false });
+
+        return this;
+    }
+
+    public editCaseOwner(): this
+    {
+        Logger.Log("Editing case owner");
+
+        this.getEditCaseOwner().click();
+
+        return this;
+    }
+
+    public canEditCaseOwner(): this
+    {
+        Logger.Log("Can edit case owner");
+
+        this.getEditCaseOwner();
+
+        return this;
+    }
+
+    public cannotEditCaseOwner(): this
+    {
+        Logger.Log("Cannot edit case owner");
+        this.getEditCaseOwner().should("not.exist");
+
+        return this;
+    }
+
+    public hasCaseOwnerReassignedBanner(): this
+    {
+        Logger.Log("Has case reassigned banner");
+
+        this.getCaseOwnerReassignedBanner().should("contain.text", "Case has been reassigned");
+
+        return this;
+    }
+
+    public hasNoCaseOwnerReassignedBanner(): this
+    {
+        Logger.Log("Has no case reassigned banner");
+
+        this.getCaseOwnerReassignedBanner().should("not.exist");
 
         return this;
     }
@@ -451,6 +514,16 @@ class CaseManagementPage {
         return cy.getByTestId("edit_Button_SFSO");
     }
 
+    private getEditCaseOwner()
+    {
+        return cy.getByTestId("edit-case-owner");
+    }
+
+    private getCaseOwnerReassignedBanner()
+    {
+        return cy.getByTestId("case-reassigned-success");
+    }
+
     private getEditIssue()
     {
         return cy.getByTestId("edit-issue");
@@ -526,18 +599,30 @@ class CaseManagementPage {
         return this;
     }
 
-    public hasConcerns(value: string): this
+    public hasConcerns(concern: string, ratingTags: Array<string>): this
     {
-        Logger.Log(`Has concerns ${value}`);
+        Logger.Log(`Has concerns ${concern}`);
 
-        cy.getByTestId(`concerns_Field`).should("contain.text", value);
+        cy.getByTestId(`concerns_Field`).should("contain.text", concern);
+
+        let concernsRow = cy.getByTestId("concerns_Field").contains(concern);
+
+        let parentRow = concernsRow.parent();
+
+        parentRow.find("td").eq(1).then((element) =>
+        {
+            ratingTags.forEach(rating =>
+            {
+                expect(element.text()).to.contain(rating);
+            });
+        });
 
         return this;
     }
 
     public hasTerritory(value: string): this
     {
-        Logger.Log(`Has territort ${value}`);
+        Logger.Log(`Has territory ${value}`);
 
         cy.getByTestId(`territory_Field`).should("contain.text", value);