+++ ConcernsCaseWork/ConcernsCaseWork/Pages/Case/CreateCase/NonConcernsCase/SelectActionOrDecision.cshtml.cs
@@ -5,6 +5,7 @@ using ConcernsCaseWork.Logging;
 using ConcernsCaseWork.Models;
 using ConcernsCaseWork.Pages.Base;
 using ConcernsCaseWork.Redis.Users;
+using ConcernsCaseWork.Service.Trusts;
 using ConcernsCaseWork.Services.Cases.Create;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -19,6 +20,7 @@ namespace ConcernsCaseWork.Pages.Case.CreateCase.NonConcernsCase;
 [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
 public class SelectActionOrDecisionPageModel : AbstractPageModel
 {
+	private readonly ITrustService _trustService;
 	private readonly IUserStateCachedService _cachedService;
 	private readonly ILogger<SelectActionOrDecisionPageModel> _logger;
 	private readonly IClaimsPrincipalHelper _claimsPrincipalHelper;
@@ -34,12 +36,14 @@ public class SelectActionOrDecisionPageModel : AbstractPageModel
 		IUserStateCachedService cachedService,
 		ILogger<SelectActionOrDecisionPageModel> logger,
 		IClaimsPrincipalHelper claimsPrincipalHelper,
-		ICreateCaseService createCaseService)
+		ICreateCaseService createCaseService,
+		ITrustService trustService)
 	{
 		_createCaseService = Guard.Against.Null(createCaseService);
 		_cachedService = Guard.Against.Null(cachedService);
 		_logger = Guard.Against.Null(logger);
 		_claimsPrincipalHelper = Guard.Against.Null(claimsPrincipalHelper);
+		_trustService = Guard.Against.Null(trustService);
 	}
 	
 	public ActionResult OnGet()
@@ -64,7 +68,12 @@ public class SelectActionOrDecisionPageModel : AbstractPageModel
 			{
 				case Options.Decision:
 					var userName = GetUserName();
-					var caseUrn = await _createCaseService.CreateNonConcernsCase(userName);
+					var state = await _cachedService.GetData(userName);
+
+					var trust = await _trustService.GetTrustByUkPrn(state.TrustUkPrn);
+					_ = trust ?? throw new NullReferenceException($"Trust information not returned for UKPRN:'{state.TrustUkPrn}', cannot continue with method execution");
+					
+					var caseUrn = await _createCaseService.CreateNonConcernsCase(userName, trust.GiasData.UkPrn, trust.GiasData.CompaniesHouseNumber);
 				
 					return Redirect($"/case/{caseUrn}/management/action/decision/addOrUpdate");